added payment

+++ exercise-2/app/models/Payment.scala
@@ -2,7 +2,7 @@ package models
 
 import play.api.libs.json.Json
 
-case class Payment (var id: Float, var status: String)
+case class Payment (var id: Long, var status: String)
 
 object Payment{
   implicit val paymentFormat = Json.format[Payment]
